Create `Picker`s for picking questions from the remaining list.
Define a type to capture the notion of a function that can give you
an item from the list, in your choice of computational context.

This made it easier to figure out how to relax the `MonadRandom` constraint
when just picking the first question from the list, but it's not actually
necessary.