Refactor to use `Deck` to store `Question`s.
Think of a deck of playing cards. You can draw from the top, draw from the
bottom, or draw at random. The same for insertions.

Expose draw and insert operations, which can be modified with locations
to target the right part of the deck.

Cut out some complexity by always drawing at random, and always inserting
at the bottom of the deck.